feat: add Discord commands for posting and support metrics
