--- README.md
@@ -1,1 +1,0 @@
-# Discord Jugalbandi AI bot
